refactor: extract LLM tab into modular component architecture
- Create shared atoms for state management (processedData, randomNumber, htmlId, feedbackSent, activeExtractTab)
- Extract LLM logic into dedicated useLlm hook
- Create reusable useFeedback hook for feedback functionality
- Refactor LlmTab component to use atoms instead of props
- Update MdrTab to use atoms for consistency
- Create LlmInteractionSection wrapper component
- Eliminate prop drilling throughout component hierarchy
- Maintain feature parity while improving code organization
